Colocation notify - bad scores refactor and fixes (#2022)
# Description
Contains various fixes around scores validation. Currently, score
validation is messy and very complicated. I tried to summarize all
checks to only 4 of them:

1. Score > 0
2. Score <= Quality, where Quality = surplus + fees.
3. SuccessProbability in range [0,1]
4. ObjectiveValue > 0, where ObjectiveValue = Quality - GasCost

(1) and (2) are checked for all scores as per CIP20. 
(3) and (4) are checked only for `RiskAdjusted` scores (scores based on
success probability used by `ScoreCalculator`), and they represent
validation of the input to the `ScoreCalculator` which demands these
checks in order for computation to be meaningful.

# Changes
- [ ] Moved out `ObjectiveValueNonPositive` and
`SuccessProbabilityOutOfRange` so that all errors from score calculator
are now boundary errors as they should be (I don't want to import
`ScoringError` into `boundary`)
- [ ] Fixed a bug: `score > quality` instead of `score >
objective_value`
- [ ] `ScoreHigherThanObjective` removed completely since this is an
internal error in score calculator and it should never happen unless
there is a bug in the code. Anyway, this is definitely not a special
case I want to handle.
- [ ] `TooHighScore` replaced with `ScoreHigherThanQuality`