Colocation Notify - failed simulation (#2013)
# Description
Implements (2) from https://github.com/cowprotocol/services/issues/1891

# Changes
- [ ] Simulation failure error (together with tx data) propagated from
the `driver` to the `solvers`.

Unfortunately, not all data can (and should) be populated to support the
legacy version. `block_number` is one of those that is needed but it is
not provided by simulators. This is the never ending issue we generally
have with doing simulations on a specific block number.

The easiest fix would be to add get_block_number() call to each web3
simulation (enso and tenderly seem to give block_number in their
responses), but it wouldn't be accurate.

At the moment I decided to skip it, since, after all, solvers can pretty
much figure out the block number on their own, there are usually 1 or 2
candidates...