Respect deadline in Baseline solver (#2040)
# Description
Baseline solver iterates through orders from `auction` and executes some
logic for each order in a synchronous way. It checks the deadline before
each order. It could happen that the new order logic begins to execute
1ms before the deadline and lasts arbitrary amount of time (500ms?) and
significantly break the deadline.

This code adds a checker if the deadline is reached in the main thread,
and if it is, ends the background work prematurely and returns whatever
was executed up until that point.

## How to test
Example:
https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=0e23fa016153883faad8b0e7148b4dfa