Notify timeout early (#2044)
# Description
Timeouts in legacy solvers are handled inside `solvers` crate and in
those cases empty solutions vector is sent to `driver`, leading to these
errors being classified in the `driver` as `empty` solutions instead of
`timeout`. The consequence of this is that legacy solvers are not
properly notified.

This PR sends the `timeout` notification directly from within `legacy`
solver so there is no ambiguity. Note that there is no way for `driver`
to know this is a `timeout`, so additional `empty` notification will be
sent, but afterwards, so it can be ignored by `legacy` solvers.