Notifies solver if their solution simulates successfully at least once (#2077)
# Description

As a solver, I am interested in automatically identifying "bad" tokens
so that they can be automatically censored. A "bad" token is one that
has some logic that makes it fail most of the times. When I receive a
simulation failure, I run an online learning procedure to try to
identify them. However, some (most?) simulation failures happen just due
to volatilty, in which case this procedure should be skipped, or
otherwise there can be false positives.
 

# Changes

This PR adds a boolean to the notification sent to solvers, stating if
the driver was able to simulate the solver solution at least once. If
so, then this is a very strong indicator that the solution was rejected
due to volatility.

I considered other options to avoid this PR:

1. Send the liquidity fetch block number in the instance.json. Since the
notification comes with a block number, then comparing the notification
block number with the liquidity fetch block number could indicate that
it simulated once at least. But this would still require a PR.
2. Same as previous, but fetch the liquidity block number from the
solver competition endpoint. However, I recall that that endpoint was
(is?) significantly behind the notification.
3. Use tenderly myself. This seems like a waste of resources just for
doing something the driver is already doing anyway.

## How to test

I did not test it since I forgot how to. In fact I would be very
surprised if I changed things in the right places. Anyway, if I failed
to do so, consider the PR as an elaborate feature request :)

Thanks!

---------

Co-authored-by: Felix Leupold <felixleupold90@gmail.com>