Rename solver_fee when referring to unsubsidised scoring fee (#2152)
# Description
Follow up in #2151 this is an attempt to rename most of the
driver/solver relevant use of "solver_fee" when meaning the
"unsubsidised fee that solvers should consider when evaluating the
object value for this order" since it's confusing when we also refer to
the fee that solvers chose e.g. for limit orders as "solver fee"

Notable, it doesn't change the field on the `order.metadata` struct, as
I'm afraid this may be used by the frontend or other dependencies.

# Changes
- [ ] Rename `solver_fee` to `scoring_fee` on `Trade`
- [ ] Cascading renames from there on
- [ ] Stop at the `order.metadata` struct (there we keep surplus_fee for
backwards compatibility reasons). It's sufficiently documented on this
struct

## How to test
Just™️  a rename 🤞 , CI still passes.

<!--
## Related Issues

Fixes #
-->

---------

Co-authored-by: narcis96 <7252787+narcis96@users.noreply.github.com>