Remove obsolete solver competition data (#2143)
# Description
Now that we collect all the competition data we care about going
forward, we can remove the obsolete fields. Changes should be backwards
compatible as `Options` will be parsed as Some if present and extra
fields that remain in the json that is stored in the DB will be ignored.
Very old auctions for which we don't store ranking or score will fail to
load (but that's acceptable IMO)

# Changes
- [ ] Make it so that we don't log calldata fields that don't exist for
non-winners.
- [ ] Remove objective as it's not recorded for runners up (and isn't
really relevant just for the winner). All we care about now is score
- [ ] Remove gasPrice and liquidityFetchedBlock as they don't exist
anymore.
- [ ] Make mandatory score and ranking non-optional
- [ ] Add e2e test for multiple solver solutions

## How to test
Run e2e test

## Related Issues

Fixes #1949