Ethflow order events (#2244)
# Description
We are trying to collect events throughout the lifecycle of all orders
for analytics and alerting reasons. It seems that we have missed native
ETH flow orders (which get indexed inside the autopilot) leading to
inconsistent lifecycles for some orders (e.g. missing creation or
cancellation events)

# Changes

- [x] insert an order event when parsing onchain order creations
- [x] insert an order event when parsing onchain order invalidations

## How to test
Adjusted the ethflow e2e test