Remove solver_fee (#2272)
# Description
Removes `solver_fee`.

I want to drop the `solver_fee` completely. It's hard to understand
whether it is a fee set by solver or it is a fee sent to solvers to do
something with it.

Also, we are currently not using subsidies (which are btw used only for
market orders that we want to deprecate) so `solver_fee` is equal to
`user_fee`.

@fleupold already touched this renaming before. After this change, only
`model::Order` will contain `solver_fee` but changing that is a major
breaking api change.

I expect shadow solver to fail again because of this. Alerter is not
deserializing this field so it should be ok.