Add block_number to order_execution (#2339)
# Description
Adds `block_number` column to db table `order_execution`.

This is to make the table reorg friendly, meaning, whenever reorg
happens, event updater will also delete all entries in this table that
are from reorged block.

This enables us to remove the delay of 64 blocks in the
`OnSettlementEventUpdater` component. Removal will be implemented in a
separate PR so that it can be easily reverted if need be.