Fix forked mainnet test (#2361)
# Description
I think we actually broke the test when we merged some of the verified
quotes logic (that's the problem with flaky test, people tend to just
ignore them when they fail)

Even locally, I'm getting a driver/solver timed out trying to place the
order. This is because the first run of the driver requires fetching all
AMMs and liquidity sources from scratch which can be a time consuming
endeavour. After that the cache is warm and we can respond fast.

# Changes
- [x] Run a dummy quote to warm up the cache in the e2e test before
placing the order.

Alternatively, I thought about making the time limit configurable and
changing it in the e2e tests but we might actually want to get a failing
test if we are unable to process quote requests in < 5s.

## How to test
By running the flaky e2e test, and seeing it now passes reliably.